Post-merge compilation fixes for stubdom
Much of this is very ugly.  Sadly that's a consequence of the stubdom
and upstream build systems and nothing much can be done about it.

However on the upside with this change we discontinue use of the
upstream `configure' script for stubdom (as it gives wrong answers),
in favour of a new xen-setup-stubdom script.

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>

committer: Ian Jackson <Ian.Jackson@eu.citrix.com>
git-commit-id: c609854fe18aab3216f28e1e5291b75970d398d5