cirrus: Do not clear vram area to 0xff when not yet allocated.
When restoring, the vram data pointer is not valid until (depending on
the version of the savefile) the guest has informed us what memory to
use.  Thus vram_ptr may be NULL.  In this patch we avoid attempting to
clear the video ram in this case.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>

committer: Ian Jackson <Ian.Jackson@eu.citrix.com>
git-commit-id: 545882583159e9cc53cde5965277967092459f37