qemu-img-xen: Fix cow creation with raw image backing files
Previously
  dd if=/dev/zero of=raw_image count=100
  ./qemu-img-xen create -b raw_image -f qcow2 cow_image
would fail because it would attempt to autodetect the format of
base_image.  With this change, it works.  However,
  ./qemu-img-xen create -b cow_image -f qcow2 cow_image_2
will now fail because we now always expect the backing file to be raw.

Thanks to Xin Li and Dongxiao Xu for the bug report.

Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>

committer: Ian Jackson <Ian.Jackson@eu.citrix.com>
git-commit-id: be598c20435f39f2d71f4b3d8d0489dd9e3af4ba