qemu xen upstream synch (Gerd): [PATCH 4/7]: no malloc failure checks.
qemu_malloc() will never return errors.  It will abort instead.
Thus no checks for failure are needed.

[This is] a series of patches for qemu-xen, making the code identical
to the xen support patches being submitted to upstream qemu.  The review
process on qemu-devel resulted in a number of fixes and cleanups in the
backend code, this is where most of the changes come from.

There are also some xenfb changes due to displaystate reorganization and
xenfb being merged in steps due to that.

Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>

committer: Ian Jackson <Ian.Jackson@eu.citrix.com>
git-commit-id: faa3347e6e7bc4137569342a191639ee8dd5d410