passthrough: use devfn instead of slots as the unit for pass-through
This is part of support for multi-function PCI devices in guests

Instead of reading a slot number from xend, read a devfn.
This and subsequent other changes will allow xend to ask
for more than one function to be inserted into a single slot -
by specifying which function of the slot should be used.

This is a minimal patch for this change. A subsequent
patch that has a lot of noise to rename slot to devfn follows.

This patch breaks compatibility with xend and corresponding
patches to xend are required.

Cc: Dexuan Cui <dexuan.cui@intel.com>
Cc: Masaki Kanno <kanno.masaki@jp.fujitsu.com>
Signed-off-by: Simon Horman <horms@verge.net.au>

[3/8; cross-compatibility issues with xen-unstable.hg]

committer: Ian Jackson <Ian.Jackson@eu.citrix.com>
git-commit-id: 7551a514dd944a5155747071135bac11deb61c43