passthrough: multi-function PCI in ACPI
This changes the interface beween qemu-xen and the ACPI portion
of hvmlader. As such it breaks compatibility with older
hvmloaders and a companion patch to hvmloader is needed.

There is a subsequent "noise" companion to this change
which changes slot to devfn as appropriate.

Cc: Dexuan Cui <dexuan.cui@intel.com>
Cc: Masaki Kanno <kanno.masaki@jp.fujitsu.com>
Signed-off-by: Simon Horman <horms@verge.net.au>

[6/8; cross-compatibility issues with xen-unstable.hg]

committer: Ian Jackson <Ian.Jackson@eu.citrix.com>
git-commit-id: a41a773ac5a8aed20222593711509d43bafaf275