pxe: Don't look for PXE ROMs when running as a DM
Searching for PXE ROMs in attached NICs is unnecessary as the ROM
that is used is included in the hvmloader binary. In the case where
there are only pass-through NICs it will cause the boot to fail
as the NICs are not known to qemu-xen at the time that this check is made.
In other words, without this change PXE boot from a pass-through
device will fail.

This has likely been broken since I changed around the pass-through
code so that devices passed-through at boot-time are handled by hot-plug.

Signed-off-by: Simon Horman <horms@verge.net.au>

committer: Ian Jackson <Ian.Jackson@eu.citrix.com>
git-commit-id: e084d4f0f1202f82db360c03c068433144d308b8